util: add analysis filter to log splitter (#17)
Adds a "Filter Log for Analysis" checkbox to the web splitter tool, and
an `-af` | `--analysis-filter` argument to the `split_log` utility. If
this option is enabled, the log lines are run through a filter and only
"useful"/"interesting" lines are emitted.

(I also added an `-na` | `--no-anonymize` argument to `split_log` to
allow anonymization to be turned off -- it's on by default. This was a
TODO in the script already, and since I was already poking around, it
made sense to kill two birds.)

The filter to determine whether a log line is useful/interesting should
essentially mirror the regex that quisquous has previously shared and
that's mentioned in the timeline guide, with one exception: I also added
comparable LosesEffect (30) lines, since the loss of a debuff may also
be interesting for timing reasons (particularly if it's removed by a
mechanic before the duration expires). However, I think reasonable minds
can certainly differ on what log lines may be useful, so I'm open to
feedback on adding to or pruning the list.

Tested against a handful of logs and working as expected, but more
testing always welcome.