util: update gen_data scripts to use xivapi library (#28)
Util cleanup project, part 2 of 3 (ref: #26)

This updates the remaining `util/gen_(data)` scripts to use the new
xivapi library. I've deleted `coinach.ts`, since we are no longer
reading SaintCoinach data directly and the file-writer function is now
in the xivapi library.

All 3 scripts were run and updated output files are included. The
changes in `effect_id.ts` and `pet_names.ts` are routine; the underlying
game data has updated since the last time these scripts were run. (These
same changes occur if you run the unmodified scripts.)

The `world_ids.ts` file has substantial changes. The original xivapi
call in that script did not include a `limit=` param in the url; as a
result, xivapi was only returning 100 records (and that's why we only
have 100 zones in the file currently). This default limit is referenced
[here](https://xivapi.com/docs/Game-Data). There are actually 861
records from the `World` [endpoint](https://xivapi.com/World), so we've
never imported the bulk of them. Most are trash/dev/test zones, but we
have been omitting a few valid public worlds.

The new xivapi library picks up all records, but the SNR here is not
ideal. Two possible options:
1. Import everything, que sera sera. But we should add the isPublic flag
so we have a way to differentiate. (This is the PR currently)
2. Use `isPublic=true` to filter the output. This is much more
manageable and adds the missing zones we want, but it does drop trash
worlds that are in the current file. (With this approach, we don't need
to set isPublic in the output, as it's implicit). I'm attaching the
output for this option in case anyone wants to run a diff.

[world_id_option_2.txt](https://github.com/OverlayPlugin/cactbot/files/13804000/world_id_option_2.txt)

I lean toward option 2, since I think that covers our use cases. If
that's the consensus, I'll do a quick fix before we land the PR.
(cc: @valarnin - I think you last worked on `gen_world_id` for
raidemulator. Any opinion?)