util: new XIVAPI CLI query tool (#41)
Adds a new XIVAPI CLI query tool to util. It's a lightweight script that
fetches endpoint data, filters it based on a user-specified filter
string, and outputs it as JSON to the console. I put an early version of
this together for looking into some of the data issues during the util
script cleanup, and it was pretty helpful. It's grown a bit since. 👀

More info on usage available by running `npm run query -- -h`.

Since it uses regex on user-supplied inputs, there's a lot of error
handling needed to capture wonky user input. I *think* I've accounted
for most of the possibilities, but open to anything I missed.

Comments/suggestions welcome!