raidboss: The Gilded Araya timeline (#50)
Thanks so much wexxlee for getting the triggers in place so quickly!

I was lucky enough to get a 9-minute pull in-game and the timeline
seemed okay. Please do check that I haven't misplaced any abilities when
accounting for multiple possibilities in a given line. Overall I feel
good about this after seeing it in-game, but it's very likely I've
missed something.

cc @wexxlee @valarnin