ci: improve release workflow (#62)
(edited after most recent commit)

When the project was moved to the OP repo, the `release.yml` workflow
was updated so that releases are no longer published in draft state.
This means that releases happen automatically once a version-bump PR is
merged; however, the release notes do not include any summary in the
title (only a version number), and the release notes will be published
instantly without edits. While the release notes can always be manually
edited by a contributor after the fact, I don't think that's ideal.

This PR adds an npm script, `npm run release` that collects the
version-bump type, a release summary, and whether the release should be
in draft, and it stores that data and runs npm version to trigger a
version bump PR (replacing the need to run `npm version`). This also
updates the `release.yml` workflow to collect these attributes and
integrate them.

(Note: Because of the changes, merging this PR will trigger the release
workflow, even though I'm not intending to actually trigger a release.
Because there's no version bump, however, the worklow will exit
gracefully without doing a release. I do plan to PR a version bump since
it's been a month and there are some i18n/bugfix PRs that have been
merged since. But I'd prefer to land this first.)