util: add new OP lines to log splitter (#92)
Proposing to add some of the new OP log lines to the log splitter
analysis function, specifically:

- `NpcYell` (266) & `BattleTalk2` (267) - possibly useful for
identifying phase jumps or mechanics
- `ActorSetPos` (271) - useful for identifying combatant movement
associated with a mechanic (esp. with forthcoming changes to
`LineCombatant`)
- `SpawnNpcExtra` (272) - initial tether/animation states for combatants
- `ActorControlExtra` (273) animation changes (although probably revisit
this if high-volume categories are added later)

`ActorSetPos` is the most "spammy" -- although I think that's probably
also fight-dependant -- so open to excluding it. But since we already
exclude `CombatantMemory` (much more "spammy"), it seems like adding
something here would be good, or we'll otherwise have to refer back to
the unfiltered log quite a bit.

Here's the impact of this change using a recent ~6m run of Asura:

| Line Type           	| Current 	| As Modified 	|
|---------------------	|---------	|-------------	|
| ALL                 	| 376     	| 498         	|
| `NpcYell`           	| 0       	| 8           	|
| `BattleTalk2`       	| 0       	| 13          	|
| `ActorSetPos`       	| 0       	| 97          	|
| `SpawnNpcExtra`     	| 0       	| 4           	|
| `ActorControlExtra` 	| 0       	| 0           	|