util: add export-all option to web splitter (#97)
Adds an "Export Entire Log" button to the web splitter. It will be
disabled if a particular fight is selected (essentially, it will toggle
states with the "Export Selected" button), and will otherwise be enabled
only if either "Anonymize Log" or "Filter Log for Analysis" is checked
(since without one of those, it would just dump the entire file,
unmodified, back to the user which is rather pointless).

Also cleans up a couple of the `canAnonymize` properties in
`netlog_defs`.

Ref #96 - this is necessary to allow a full log file to be anonymized
and exported (without being split), in case the splitter is not properly
detecting encounters.