ci/raidboss: add timeline linter & lint timeline files (#106)
Per discussion in #96 and with @JLGarber , this adds some timeline
linting to `mocha` (which necessitated actually fixing the lint.... and
there was a *lot*).

Sorry for the sheer volume of files changed, but the vast majority are
just one or a couple of lines with routine changes like excess
whitespace removal, reordering syncs that were out of order, removing
duplicates, fixing a few obvious typos, etc. There were, however, a
couple of Omega timelines that had branching/jumps that were out of sync
with our current practice, so I revised them to be in chron order rather
than just ignoring sync order for the whole file.

The scope of the linting is covered in the updated `TImelineGuide.md`,
but at a high level, the major changes are:
- Enforcing all syncs to be in chronological order.
- Note: this can be disabled for a particular block of entries by using
`#cactbot-timeline-lint-disable-sync-order`, followed by
`#cactbot-timeline-lint-enable-sync-order`. This should ideally be
limited to situations where the timeline includes branching with some
fake lookaheads -- basically, where it's easier for readability to have
a couple blocks of overlapping times.
- Requiring keywords (`duration`, `window`, etc.) to come after the
`[LogLine]` and `NetRegex` values on sync lines.
- Requiring keywords (if used) to be in the following order: `duration`,
`window`, `jump`/`forcejump`.
- Requiring any use of `window` to specify both a lookbehind and
lookahead (e.g. `[number],[number]`), rather than just a single param as
was often done previously.

Comments/feedback welcome.