ci: add check for "point-zero" timeline lint (#110)
Addresses @JLGarber 's comment in #106 about linting for "xx.0" float
parameters for keyword values in timelines that should just be ints.

(Also fixes an inadvertent bug that was causing a few mis-ordered
keywords to not be linted properly.)

Given that this will again touch a decent number of old timelines
(although not as many as #106), I've temporarily added a control to
limit the lint scope, and this PR only includes changes to EW+ShB
timelines (+sophia_ex and zurvan_ex due to sync-files). Will do
follow-on PRs for SB (there are a lot) and ARR+HW before the expac
control filter is removed.