raidboss: initial ex1 (#188)
Edit: This should be mostly done now. Triggers are complete, timeline
looks good (huge thanks to @valarnin). There's some bare-bones oopsy
that really only covers the first couple minutes of the fight -- I may
try to add more tomorrow, but that might also be a follow-up PR so we
can get this out.

I've tested against a number of logs in raidemulator with vod, and
things generally look good. But this would certainly benefit from more
in-game testing.