build: Bump version to 0.32.1 (#193)
Pushing ex1 support out. Releasing in draft so the release notes can be
edited to (a) note that ex2 is next, followed by other content, and (b)
include a link to the content coverage page. Hopefully this should quell
some of the questions in the ACT discord about DT content support...