raidboss: ex2 timeline+triggers (#216)
This is an initial pass at ex2. Original timeline thanks to @valarnin
(this PR replaces #192), with some adjustments/cleanup from me, but I
think the timeline is largely good.

Triggers have been tested against raidemulator but still need some
testing in game and may need a little cleanup in some spots (marked as
draft/wip for now, but I should be able to get back in tonight). There's
a to-do list at the top for some mechs that do not yet have triggers.
I'm going to keep poking at them tonight, but it may make sense to merge
as-is and do with follow-up PRs.

---------

Co-authored-by: valarnin <valarnin@gmail.com>