util: switch to xivapi beta (#171)
This is an initial pass at updating the xivapi lib and the gen_* util
scripts that use it to take advantage of the new xivapi beta API. The
line count is running a little high, but the vast majority of the
changes are due to the API schema changes. The only notable code
refactoring is in `util/xivapi.ts`, and mostly relates to the way
looping through result sets now has to be done due to the absence of
pagination properties.

Running `npm run generate -- -f all` also updated some of the resource
files. The diff seems to be mostly new things added in 6.55, along with
a small bug fix in the way that synthetic zone info is handled. (NB: We
can probably kill off the support for the pre-rework dungeons now that
cn/ko is caught up, but I'd rather do that in a separate PR.)

~~Marking this as WIP for now so I can look with fresh eyes tomorrow and
probably tidy a few more things up. But feedback welcome.~~ Should be
ready for review.