raidboss: fix R1N Black Cat Crossing (#265)
Right now `R1N Black Cat Crossing` always outputs `Intercards => Cards`
regardless of the proper order. This seems to be happening for two
reasons. The `heading` field for `StartsUsingExtra` line is listed as
the `firstUnknownField`, which means it is never being captured by
regex. Second, and unrelated, the output calls in the trigger are
inverted.

@valarnin - not sure if there was a reason why the `heading` field was
originally treated as optional (e.g. if there is ever a circumstance
that field won't be populated from OP), or whether that was just a
copy-paste thing?