resources: 7.05 updates + files (#289)
Resource file updates and content skeletons for Patch 7.05.  
This maintains parity with the current naming convention, but can be
updated later if #272 is merged (sidenote: if so, the oopsy trigger
files should also be renamed for both normal and savage.)
This also updates the current `effect_id.ts` file notwithstanding the
changes proposed in #286.