jobs: remove valuescale setting (#292)
https://github.com/OverlayPlugin/cactbot/blob/1be5017fe1233268ab6dd10cf8d206f2921508c7/resources/timerbox.ts#L488

The valuescale setting seems to be used as a divider before showing the
timerbox text (i.e the remaining time), so translate the plain time to
be how many GCDs remaining if we set it to the current GCD length.
But it looks like most of the time, instead of seeing the text, we are
just watching the box showing larger -> prepare to press the
corresponding button. @Echoring would explain it more!