util: splitter/anonymizer fix (#321)
On more recent versions of NodeJS, it will complain about using a named
import from lodash (specifically, `isEqual` which is used in splitter &
anonymizer for typeguards) because it is a CommonJS module. This fix
allows `split_log` to be run from the CLI (which is currently broken).

This also updates the logdef for 0x1B lines. In R4S, one of the
uncaptured params contains a duplicate of the actorId, which causes
anonymizer to throw uncaught-player errors. This adds that field as a
`possiblePlayerId`, meaning it will be silently anonymized if present
but not included as a regex capture group.