raidboss: savage 4th floor (#311)
Timeline is done, barring a mistake somewhere. Triggers are still wip -
p1 is done, but doesn't have much p2 yet. Planning to get at least a
little more p2 in there before this gets merged.

This looks fine in raid emulator, but it could probably use some in-game
testing if anyone is able to. Worst case, we can merge and I can in-game
test and adjust if needed on Tuesday.

---------

Co-authored-by: valarnin <6119598+valarnin@users.noreply.github.com>