raidemulator: Fix bug with double seek in raidemulator (#329)
This is the cause of raidemulator not seeking to the start of the fight
properly when going from a newer encounter to an older encounter.

Basically, the previous encounter's timestamp was held when changing
encounters, which causes the initial perspective selection to seek to
that timestamp. If the previous encounter was before the new encounter,
that's not a big problem since the seek would be capped to the start of
the encounter.

But if the previous encounter's timestamp was later than the current
encounter, the seek would succeed and be past the end of the newly
loaded encounter.