raidboss: Additional M/R2S triggers (#356)
Finally getting around to PRing some additional triggers:
* Revises the 'Stored: ' partner/spread call to an `infoText` and
suppresses it for the first two occurrences (since Tempting Twist/Honey
Beeline happen immediately after).
* Adds an `infoText` with the subsequent positioning + partner/spread at
the beginning of Tempting Twist and Honey Beeline, followed by an
explicit `alertText` with that info when the first part of the mechanic
resolves.
* Lists some possible TODOs for beat1 & beat2.
* Adds defamation/in/towers calls during beat3. (NB: I had this mostly
worked up before #349, but if people prefer that approach + some added
handling for defamations, I can remove it from this PR before merging.)
* Adds callouts for "curtain call" (aka Beeloved Venom). On this, I
added a 'Marge w/ X` alert for the active player, as well as info alerts
for other merges that should happen not involving the active player. On
the latter, I could see some disagreement about whether that info is
useful or not, so happy to adjust/remove that if folks don't like.

(NB2: This PR doesn't address/replace anything in #339 or #343 -- I had
in my notes to do things similarly to those, but since there are already
PRs, I think it makes sense to work with those.)

This looks fine in vod/RE, but probably could use a proofread and some
in-game testing. I may be able to test a little later tonight, but would
be great if others can poke at it too.