raidboss: r3s Barbarous Barrage Uptime Knockback (#316)
The first tower is generally the hardest to get the angles correct due
to the greatest distance of the knockback and closeness to the center of
the knockback origin. The first tower dodge is also most significant for
melee uptime. This trigger defaults to dodging the first two towers.

A 15s delay will dodge all three towers. This is very tight and if
missing the first tower, it will almost guarantee missing the next two
towers. A 12s delay dodges the first two towers. This is lenient, but
too late and you likely will end up with a damage down from Murderous
Mist. A 9s delay will dodge only the first tower. Lateness on this could
more easily lead to a death or missed tower due to angle of knockback.

Since missed tower soaks result in a stacking raidwide DoT debuff. I
think defaulting to 12s is the safest as it seems most likely to
recover.

Any tower explosions that are dodged require moving to the next tower.
Sprint is unnecessary for first two, but may be helpful for longer
casts. If the middle tower is dodged, you likely need some sort of
jump/teleport ability to avoid getting hit by murderous mist as you only
have 2s to get in position rather than the 3s between tower explosions.

Another thing to note, the explosion knockback happens before the tower
animation disappears, and so players would need to still learn the
timing when using this strategy since they should move before the tower
disappears for more time to move to the next tower. A move trigger could
be beneficial for this, but could unnecessarily add noise.

There was discussion over whether this should be a user trigger or not.
It may take on in PF as it grants melee uptime, and removes a tight
position requirement for the first tower without too strict of timing.