raidboss: Initial Deadwalk package (#360)
Should be largely self-explanatory. Oopsy has not been tested in-game.

I'm not sure what we really want to do about boss 1. It's very much "use
your eyes" and not a lot of value from triggers.

Boss 2 we could probably map out the patterns based on that one
infographic going around, but that's a lot of extra work, and I kind of
want to get the timelines/oopsy in place for now so people have it
available.