resources: add NPC-source+target effects to analysis filter (#387)
Small fix to the log splitter's analysis filter, based on something I
noticed while doing the A-rank triggers yesterday. Right now, the
analysis filter does not include `GainsEffect`/`LosesEffect` lines that
originate from an NPC (4.{7}) that are applied to an NPC, e.g.,
self-applied boss status effects. This PR fixes that.

Extremely small overall impact to the number of lines that pass the
filter, but they are important ones.