raidboss: add support for live countdowns (#358)
Something that's been lingering on my to do list for a while, but some
of the conversations on PRs today made me finally get to it. This adds
an optional trigger property `countdownSeconds:`, which supports either
a number or a `(data, matches) =>` function. If specified, it will
append a live countdown (floating point, tenths place) to any popup text
generated by that trigger. (As currently written, the countdown portion
of the text will not be output via tts, because that gets pretty messy.)

I added a demo of this to the countdown start trigger in Summerford
Farms, so this can be easily tested with the usual `/cd 5` there.

The only significant judgment call (which I'm very open to revisiting)
was how to handle potential interactions between `countdownSeconds` and
`durationSeconds` (either specified on trigger, in user config, or the
default). As written:
* If the trigger's duration equals or exceeds the countdown, the
countdown will simply stop at `(0.0)` and will continue to be displayed
until the duration ends and the popup text disappears.
* If the trigger's countdown exceeds the duration, the countdown value
will override the duration (default or specified) and keep the popup
text displayed until the countdown reaches 0.0. This seemed to make the
most sense to me, but again, very open to revisiting.

This may not be something that gets frequent use, but I think there are
use cases where a callout needs to be resolved by a particular time
(whether tied to a timeline entry or not), and without an obvious visual
cue/castbar in game. So this could provide additional info in certain
circumstances to players about when to do something (like, say, merge
debuffs for a curtain call mechanic...)

Wanted to get this up for input, but if folks like the idea, I can
update the docs to reflect this capability once merged.