oopsy: refactor (fix) party tracking (#447)
In prep for some feature improvements I'm hoping to make in oopsy, there
were some bugs in party state tracking that need to be addressed first.

Because `oopsy_viewer` reads raw log lines to (re)assemble oopsy reports
and isn't a true emulator, oopsy relies entirely on `PartyList` lines
for its understanding of party composition, and simulates `PartyChanged`
events based on that. (Longer term, it'd be great to refactor the viewer
and switch all of oopsy to event handling so this wouldn't be necessary,
but that's a larger scope than I want to take on right now.) But the
current implementation regularly generates errors in the ACT OP log
about not being able to locate player ids (including sometimes for the
actual current player), and in general, oopsy's party-tracking has a few
problems:
* It assumes that `AddedCombatant` lines will always precede `PartyList`
lines -- and I've got plenty of logs where that's not the case, and as a
result, the party list is incomplete.
* If there is no `PartyList` line after a zone change (e.g. solo
instance, duty support, etc.), the partyTracker is cleared but is not
updated to include the current player, which can cause a variety of
issues (like `data.party` not containing the current user).

This is a small refactor of the logic to address these issues. I've
tested this fairly extensively with console logging and with some WIP
triggers (related to the feature that I'm working on), but more eyes
always good.