raidboss: variable-length arrow substitution in tts (#465)
As discussed in #458 and #464.

Right now, tts will strip an arrow (`<=`, `<-`, `=>`, or `->`) at the
beginning or end of an output string so it does not get spoken. Arrows
in the middle of a string are spoken as `then` (translated as needed).

This allows the beginning/end arrows to be variable length (e.g., `<==`
or `<---`) and still be stripped, so we don't end up with partial
substitution weirdness (e.g., tts speaking `equals` when `<==` is
stripped to `=`).