raidboss/oopsy: add Everkeep (normal) package (#438)
Tested in emulator and in game.

This re-uses a lot of the tile calculation logic I did for EverkeepEx,
but determining/calling safe spots is handled a little differently
because the forged track patterns are both fixed and simpler. For
calling Forged Track, it will call a single inner safe tile regardless
of how many tiles are actually safe, with priority given to the inner
N/S tiles. I *think* this probably fine, but given the recent discussion
on #429 & #431, I wanted to mention that. If folks prefer a different
priority order, feel free to PR a config option, but I think this is
fine for an initial pass at least.