oopsy: fix lower-case player id (#472)
Closes #467.

This fixes the extraneous '???' entry that was showing up in oopsy
mistake reporting. The fix to party list handling in #447 ensures that
the current player's id will always be in the party list. However,
oopsy's identification of the current player can come from
`onPlayerChangedEvent` as well as `ChangePrimaryPlayer` lines, and in
the former case, the player id was in lowercase, causing an extraneous
entry in the player list when the current player was then (re-)added (in
uppercase) from `PartyList`/`AddCombatant` lines. This PR ensures that
the current player id will always be in uppercase (and does so in
several places to future-proof `PlayerStateTracker` and `DamageTracker`
if those classes get used elsewhere later).

Massive thanks to @xiashtra who did the debugging, identified the cause,
and tested the fix.

Co-authored-by: xiashtra <91220277+xiashtra@users.noreply.github.com>