raidboss: r4s strategy options for NA strats (#453)
This series includes several strategy options for the m4s fight that fit
the
standard NA PF strategies.

- **raidboss: add support for DN strategy for Ion Cluster in m4s**
- **raidboss: add support for DN strategy during Witch Hunt in m4s**
- **raidboss: add uptime sunrise sabaath callouts in m4s**

These are based off the shabin pastebin and hector guide, with the
uptime
waymarks strategy for sunrise sabaath.

I made all of these configuration options, so that existing default
callouts
which do not use a specific strategy are not affected. I tried to put
good
brief explanations, and used the standard strat names I see going
around.

I've tested variations of this code in my runs, though it did not use a
strategy option at the time. Thus, this exact code is not yet battle
tested
outside of the simulator.

---------

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>