raidboss: add Mica the Magical Mu (worldboss: Living Memory) (#468)
I used a script to generate and verify the card patterns based on the
many logs I have. Tested what logs I have videos for and in game.

I could not decide on a nice way to call the combo versions of in/out
with the line aoes from the hoops so I just took that out.