raidboss: Basic FRU P1 triggers (#516)
This is partially based on an experimental set of triggers @valarnin
tossed in the ACT Discord server yesterday or the day before. I wasn't
able to get the Blasting Zone safe area triggers working, but apart from
that and the tank buster trigger, everything in here is stuff I was able
to test tonight and confirm as working. The timing for delays and
durations is all stuff that fits relatively closely to match mechanics.
(Castbars in phase 1 are a little longer than we're used to.)