raidboss: Add Interphos triggers/timeline (#542)
Apologies for the lateness on this one, there was a lot of fiddliness
getting the timeline to work, and then in-game testing for it is kinda
painful. It was tough to find motivation at multiple points while
assembling this.

The timeline isn't perfect but it's definitely Good Enough. Oopsy will
be forthcoming Eventually.