raidboss: FRU P3 - add swap reminders to Apoc stacks (#580)
Small tweak to the P3 Apoc stack triggers. If the player takes their
first stack on a different side of the boss than they were on when
debuffs come out, that swap is detected, and the player will receive a
reminder to stay swapped for the second (post-Apoc) and third (post-jump
+ kb) stacks.

No changes to timing, just outputs.  Tested in emulator and in game.