raidboss: tweak timing/output for FRU P1 Fall of Faith (#578)
Suggesting some slight tweaks for the P1 Fall of Faith trigger. This PR:
* extends the duration of the full sequence alert to 12.2 seconds (right
now, it disappears as the 2nd cleave is going off; this will now keep it
active until right before the 4th cleave resolves)
* includes the target of the first three tethers in their alerts, and
raises the output to `alertText` if the tether is on the user for added
awareness

Happy to keep this as a personal trigger if no one else finds it useful.