raidboss: FRU p4 triggers (#601)
~~Marking this as WIP for now because I want to get it posted for
feedback but may not be able to get back to it until next week. ~~

Working versions of these triggers have been tested quite a bit in game,
but I changed a few things before PRing so I'm planning to re-test in
game next week. (Probably also some typos in the comments, but I need to
look again with fresh eyes.)